# pkg/trace/agent: replace custom UTF-8 validation in normalization with std lib