### What does this PR do?

Replaces custom UTF8 validation/replacement with the std library one.

Additionally, the span `metrics` map keys are also added to the UTF8
normalizer.