@ichinaski 2020-10-20
`benchstat` results for both `BenchmarkToUTF8` and `BenchmarkNormalization`:

```
name                  old time/op    new time/op    delta
Normalization-8          765ns ± 2%     710ns ± 7%  -7.31%  (p=0.000 n=18+20)
ToUTF8/valid_ascii-8    91.6ns ± 6%    92.1ns ± 6%    ~     (p=0.525 n=20+20)
ToUTF8/valid_utf8-8      586ns ± 6%     570ns ± 6%  -2.61%  (p=0.001 n=20+20)
ToUTF8/invalid-8         114ns ± 4%     114ns ± 8%    ~     (p=0.839 n=19+20)

name                  old alloc/op   new alloc/op   delta
Normalization-8           952B ± 0%      952B ± 0%    ~     (all equal)
ToUTF8/valid_ascii-8     32.0B ± 0%     32.0B ± 0%    ~     (all equal)
ToUTF8/valid_utf8-8       192B ± 0%      192B ± 0%    ~     (all equal)
ToUTF8/invalid-8         20.0B ± 0%     20.0B ± 0%    ~     (all equal)

name                  old allocs/op  new allocs/op  delta
Normalization-8           10.0 ± 0%      10.0 ± 0%    ~     (all equal)
ToUTF8/valid_ascii-8      2.00 ± 0%      2.00 ± 0%    ~     (all equal)
ToUTF8/valid_utf8-8       5.00 ± 0%      5.00 ± 0%    ~     (all equal)
ToUTF8/invalid-8          2.00 ± 0%      2.00 ± 0%    ~     (all equal)
```