@ichinaski 2020-10-21
I rerun benchmarks after the update:

```
name              old time/op    new time/op    delta
Normalization-8      870ns ±23%     690ns ± 5%  -20.75%  (p=0.000 n=20+17)
ToUTF8/ascii-8       104ns ±18%      91ns ± 3%  -12.24%  (p=0.000 n=20+20)
ToUTF8/utf8-8        680ns ±31%     569ns ± 3%  -16.42%  (p=0.000 n=19+19)
ToUTF8/invalid-8     132ns ±30%     112ns ± 2%  -14.79%  (p=0.000 n=19+19)

name              old alloc/op   new alloc/op   delta
Normalization-8       952B ± 0%      952B ± 0%     ~     (all equal)
ToUTF8/ascii-8       32.0B ± 0%     32.0B ± 0%     ~     (all equal)
ToUTF8/utf8-8         192B ± 0%      192B ± 0%     ~     (all equal)
ToUTF8/invalid-8     20.0B ± 0%     20.0B ± 0%     ~     (all equal)

name              old allocs/op  new allocs/op  delta
Normalization-8       10.0 ± 0%      10.0 ± 0%     ~     (all equal)
ToUTF8/ascii-8        2.00 ± 0%      2.00 ± 0%     ~     (all equal)
ToUTF8/utf8-8         5.00 ± 0%      5.00 ± 0%     ~     (all equal)
ToUTF8/invalid-8      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
```