@gbbr 2020-10-21
That doesn't look right, I don't think you're running the comparison correctly. The command is:
```
benchstat -delta-test none old.txt new.txt
```
I believe in your case you've reversed the two so the "new" time is actually the old time and there is some performance loss. When I do it, I get:
```
name              old time/op    new time/op    delta
ToUTF8/ascii-8      91.7ns ± 0%    97.8ns ± 0%   +6.65%
ToUTF8/utf8-8        690ns ± 0%     676ns ± 0%   -2.03%
ToUTF8/invalid-8     121ns ± 0%     116ns ± 0%   -4.13%
Normalization-8      720ns ± 0%     803ns ± 0%  +11.53%

name              old alloc/op   new alloc/op   delta
ToUTF8/ascii-8       32.0B ± 0%     32.0B ± 0%    0.00%
ToUTF8/utf8-8         192B ± 0%      192B ± 0%    0.00%
ToUTF8/invalid-8     20.0B ± 0%     20.0B ± 0%    0.00%
Normalization-8       952B ± 0%      952B ± 0%    0.00%

name              old allocs/op  new allocs/op  delta
ToUTF8/ascii-8        2.00 ± 0%      2.00 ± 0%    0.00%
ToUTF8/utf8-8         5.00 ± 0%      5.00 ± 0%    0.00%
ToUTF8/invalid-8      2.00 ± 0%      2.00 ± 0%    0.00%
Normalization-8       10.0 ± 0%      10.0 ± 0%    0.00%
```