@ichinaski 2020-10-21
Ah good catch @gbbr, my command was indeed wrong. I was running `benchstat new.txt old.txt` 👍 

These would be the results with 20 runs of each version:
```
benchstat -delta-test none old.txt new.txt                                                                                                                                                                                               15:59:52
name              old time/op    new time/op    delta
Normalization-8      705ns ± 8%     868ns ±22%  +23.12%
ToUTF8/ascii-8      92.2ns ± 4%   100.1ns ± 9%   +8.57%
ToUTF8/utf8-8        581ns ± 6%     654ns ±12%  +12.69%
ToUTF8/invalid-8     115ns ± 5%     125ns ± 9%   +9.13%

name              old alloc/op   new alloc/op   delta
Normalization-8       952B ± 0%      952B ± 0%    0.00%
ToUTF8/ascii-8       32.0B ± 0%     32.0B ± 0%    0.00%
ToUTF8/utf8-8         192B ± 0%      192B ± 0%    0.00%
ToUTF8/invalid-8     20.0B ± 0%     20.0B ± 0%    0.00%

name              old allocs/op  new allocs/op  delta
Normalization-8       10.0 ± 0%      10.0 ± 0%    0.00%
ToUTF8/ascii-8        2.00 ± 0%      2.00 ± 0%    0.00%
ToUTF8/utf8-8         5.00 ± 0%      5.00 ± 0%    0.00%
ToUTF8/invalid-8      2.00 ± 0%      2.00 ± 0%    0.00%
```