@gbbr 2020-10-21
23% on top is a lot. I don't think we can afford that. Let's discuss offline and reassess.