@gbbr 2020-10-29
Closing since this was resolved by https://github.com/DataDog/datadog-agent/pull/6645